GetAJob is called if bot becomes idle
This is because there might be a job waiting to go but the flag
isn't set on the bot. This happens when the second job is uploaded
before the GetAJob call is made which grabs the first job. The flag
is cleared on the bot, but the second job is still available so we
call GetAJob just in case there's another.